refactor(iac-scan): restructure classes: scan result, scan output
