+++ ggshield/iac/output/iac_output_handler.py
@@ -4,7 +4,9 @@ from typing import Optional
 import click
 
 from ggshield.core.errors import ExitCode
-from ggshield.iac.iac_scan_collection import IaCScanCollection
+from ggshield.iac.collection.iac_diff_scan_collection import IaCDiffScanCollection
+from ggshield.iac.collection.iac_path_scan_collection import IaCPathScanCollection
+from ggshield.iac.collection.iac_scan_collection import IaCScanCollection
 
 
 class IaCOutputHandler(ABC):
@@ -19,22 +21,26 @@ class IaCOutputHandler(ABC):
         self.verbose = verbose
         self.output = output
 
-    def process_scan(self, scan: IaCScanCollection) -> ExitCode:
+    def process_scan(self, scan: IaCPathScanCollection) -> ExitCode:
         """Process a scan collection, write the report to :attr:`self.output`
 
         :param scan: The scan collection to process
         :return: The exit code
         """
         text = self._process_scan_impl(scan)
-        if self.output:
-            with open(self.output, "w+") as f:
-                f.write(text)
-        else:
-            click.echo(text)
-        return self._get_exit_code(scan)
+        return self._handle_process_scan_result(scan, text)
+
+    def process_diff_scan(self, scan: IaCDiffScanCollection) -> ExitCode:
+        """Process a diff scan collection, write the report to :attr:`self.output`
+
+        :param scan: The scan collection to process
+        :return: The exit code
+        """
+        text = self._process_diff_scan_impl(scan)
+        return self._handle_process_scan_result(scan, text)
 
     @abstractmethod
-    def _process_scan_impl(self, scan: IaCScanCollection) -> str:
+    def _process_scan_impl(self, scan: IaCPathScanCollection) -> str:
         """Implementation of scan processing,
         called by :meth:`OutputHandler.process_scan`
 
@@ -45,9 +51,31 @@ class IaCOutputHandler(ABC):
         """
         raise NotImplementedError()
 
+    @abstractmethod
+    def _process_diff_scan_impl(self, scan: IaCDiffScanCollection) -> str:
+        """Implementation of diff scan processing,
+        called by :meth:`OutputHandler.process_diff_scan`
+
+        Must return a string for the report.
+
+        :param scan: The scan collection to process
+        :return: The content
+        """
+        raise NotImplementedError()
+
     def _get_exit_code(self, scan: IaCScanCollection) -> ExitCode:
-        if scan.result is None:
+        if scan.result is None or scan.type == "unknown":
             return ExitCode.UNEXPECTED_ERROR
-        if scan.result.entities_with_incidents:
+        if scan.has_results:
             return ExitCode.SCAN_FOUND_PROBLEMS
         return ExitCode.SUCCESS
+
+    def _handle_process_scan_result(
+        self, scan: IaCScanCollection, text: str
+    ) -> ExitCode:
+        if self.output:
+            with open(self.output, "w+") as f:
+                f.write(text)
+        else:
+            click.echo(text)
+        return self._get_exit_code(scan)
